refactor: create NpmInstaller
