+++ Cargo.toml
@@ -54,7 +54,7 @@ deno_bench_util = { version = "0.179.0", path = "./bench_util" }
 deno_config = { version = "=0.43.0", features = ["workspace", "sync"] }
 deno_lockfile = "=0.24.0"
 deno_media_type = { version = "0.2.3", features = ["module_specifier"] }
-deno_npm = "=0.27.0"
+deno_npm = "=0.27.1"
 deno_path_util = "=0.3.0"
 deno_permissions = { version = "0.44.0", path = "./runtime/permissions" }
 deno_runtime = { version = "0.193.0", path = "./runtime" }
