+++ cli/npm/installer/common/lifecycle_scripts.rs
