+++ cli/npm/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
 mod byonm;
+pub mod installer;
 mod managed;
 mod permission_checker;
 
@@ -30,7 +31,7 @@ pub use self::byonm::CliByonmNpmResolverCreateOptions;
 pub use self::managed::CliManagedNpmResolverCreateOptions;
 pub use self::managed::CliNpmResolverManagedSnapshotOption;
 pub use self::managed::ManagedCliNpmResolver;
-pub use self::managed::PackageCaching;
+pub use self::managed::NpmResolutionInitializer;
 pub use self::managed::ResolvePkgFolderFromDenoModuleError;
 pub use self::permission_checker::NpmRegistryReadPermissionChecker;
 pub use self::permission_checker::NpmRegistryReadPermissionCheckerMode;
@@ -109,25 +110,13 @@ pub enum CliNpmResolverCreateOptions {
   Byonm(CliByonmNpmResolverCreateOptions),
 }
 
-pub async fn create_cli_npm_resolver_for_lsp(
+pub fn create_cli_npm_resolver(
   options: CliNpmResolverCreateOptions,
 ) -> Arc<dyn CliNpmResolver> {
   use CliNpmResolverCreateOptions::*;
   match options {
-    Managed(options) => {
-      managed::create_managed_npm_resolver_for_lsp(options).await
-    }
+    Managed(options) => managed::create_managed_npm_resolver(options),
     Byonm(options) => Arc::new(ByonmNpmResolver::new(options)),
-  }
-}
-
-pub async fn create_cli_npm_resolver(
-  options: CliNpmResolverCreateOptions,
-) -> Result<Arc<dyn CliNpmResolver>, AnyError> {
-  use CliNpmResolverCreateOptions::*;
-  match options {
-    Managed(options) => managed::create_managed_npm_resolver(options).await,
-    Byonm(options) => Ok(Arc::new(ByonmNpmResolver::new(options))),
   }
 }
