+++ cli/tools/repl/session.rs
@@ -45,6 +45,7 @@ use crate::args::CliOptions;
 use crate::cdp;
 use crate::colors;
 use crate::lsp::ReplLanguageServer;
+use crate::npm::installer::NpmInstaller;
 use crate::npm::CliNpmResolver;
 use crate::resolver::CliResolver;
 use crate::tools::test::report_tests;
@@ -181,7 +182,7 @@ struct ReplJsxState {
 }
 
 pub struct ReplSession {
-  npm_resolver: Arc<dyn CliNpmResolver>,
+  npm_installer: Option<Arc<NpmInstaller>>,
   resolver: Arc<CliResolver>,
   pub worker: MainWorker,
   session: LocalInspectorSession,
@@ -200,7 +201,7 @@ pub struct ReplSession {
 impl ReplSession {
   pub async fn initialize(
     cli_options: &CliOptions,
-    npm_resolver: Arc<dyn CliNpmResolver>,
+    npm_installer: Option<Arc<NpmInstaller>>,
     resolver: Arc<CliResolver>,
     mut worker: MainWorker,
     main_module: ModuleSpecifier,
@@ -265,7 +266,7 @@ impl ReplSession {
       )?;
     let experimental_decorators = transpile_options.use_ts_decorators;
     let mut repl_session = ReplSession {
-      npm_resolver,
+      npm_installer,
       resolver,
       worker,
       session,
@@ -704,8 +705,8 @@ impl ReplSession {
     &mut self,
     program: &swc_ast::Program,
   ) -> Result<(), AnyError> {
-    let Some(npm_resolver) = self.npm_resolver.as_managed() else {
-      return Ok(()); // don't auto-install for byonm
+    let Some(npm_installer) = &self.npm_installer else {
+      return Ok(());
     };
 
     let mut collector = ImportCollector::new();
@@ -737,13 +738,13 @@ impl ReplSession {
     let has_node_specifier =
       resolved_imports.iter().any(|url| url.scheme() == "node");
     if !npm_imports.is_empty() || has_node_specifier {
-      npm_resolver
+      npm_installer
         .add_and_cache_package_reqs(&npm_imports)
         .await?;
 
       // prevent messages in the repl about @types/node not being cached
       if has_node_specifier {
-        npm_resolver.inject_synthetic_types_node_package().await?;
+        npm_installer.inject_synthetic_types_node_package().await?;
       }
     }
     Ok(())
